requirements.txt
langchain>=1.0.0
langchain-openai>=1.0.0
langgraph>=1.0.0
sentence-transformers>=3.0.0
numpy>=1.26.0
bm25s>=0.2.0
python-dotenv>=1.0.0
